requirements.txt
fastapi
uvicorn
chromadb
google-genai
pypdf2
streamlit
python-multipart
pydantic
requests